[REFACTOR] 커스텀 응답 dto common으로 이동 #2
