+++ src/main/java/com/hackathon/org/common/dto/ApiResponseDTO.java
@@ -1,4 +1,4 @@
-package com.hackathon.org.controller.dto;
+package com.hackathon.org.common.dto;
 
 import com.hackathon.org.common.status.ErrorStatus;
 import com.hackathon.org.common.status.SuccessStatus;
