+++ src/main/java/com/hackathon/org/controller/SampleController.java
@@ -1,7 +1,7 @@
 package com.hackathon.org.controller;
 
 import com.hackathon.org.common.status.SuccessStatus;
-import com.hackathon.org.controller.dto.ApiResponseDTO;
+import com.hackathon.org.common.dto.ApiResponseDTO;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
